[dev] update to containers 1.0

+++ src/core/HVar.ml
@@ -20,7 +20,7 @@ let cast v ~ty = {v with ty; }
 let update_ty v ~f = {v with ty=f v.ty; }
 
 let compare cmp a b =
-  let c = CCOrd.int_ a.id b.id  in
+  let c = CCOrd.int a.id b.id  in
   if c<>0 then c else cmp a.ty b.ty
 let equal eq a b = a.id = b.id && eq a.ty b.ty
 let hash a = Hash.int a.id
